refactor: walk-journals -> journals, walk-dogs -> journals-dogs로 테이블명 변경
