+++ backend/src/entity/walk-log-photos.entity.ts
@@ -1,10 +1,10 @@
 import { Column, Entity, JoinColumn, ManyToOne, PrimaryColumn } from 'typeorm';
-import { WalkJournals } from '../journals/walk-journals.entity';
+import { Journals } from '../journals/journals.entity';
 
 @Entity('walk_log_photos')
 export class WalkLogPhotos {
     @PrimaryColumn({ name: 'journal_id' })
-    @ManyToOne(() => WalkJournals, (walkJournals) => walkJournals.id)
+    @ManyToOne(() => Journals, (walkJournals) => walkJournals.id)
     @JoinColumn({ name: 'journal_id' })
     journalId: number;
