+++ backend/src/journals/journals.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from 'src/common/database/database.module';
 import { JournalsController } from './journals.controller';
+import { Journals } from './journals.entity';
+import { WalkJournalsRepository } from './journals.repository';
 import { JournalsService } from './journals.service';
-import { WalkJournals } from './walk-journals.entity';
-import { WalkJournalsRepository } from './walk-journals.repository';
 
 @Module({
-    imports: [DatabaseModule.forFeature([WalkJournals])],
+    imports: [DatabaseModule.forFeature([Journals])],
     providers: [JournalsService, WalkJournalsRepository],
     controllers: [JournalsController],
 })
